Name Changes
-Name and formatting changes
-Add userid to result dictionary no matter what
