+++ pyracf/__init__.py
@@ -2,8 +2,8 @@
 from .access.access_admin import AccessAdmin
 from .common.add_operation_error import AddOperationError
 from .common.alter_operation_error import AlterOperationError
+from .common.downstream_fatal_error import DownstreamFatalError
 from .common.security_request_error import SecurityRequestError
-from .common.security_response_error import SecurityResponseError
 from .common.segment_error import SegmentError
 from .common.segment_trait_error import SegmentTraitError
 from .common.userid_error import UserIdError
@@ -11,6 +11,6 @@ from .connection.connection_admin import ConnectionAdmin
 from .data_set.data_set_admin import DataSetAdmin
 from .group.group_admin import GroupAdmin
 from .resource.resource_admin import ResourceAdmin
-from .scripts.install_precheck import define_precheck_profile
+from .scripts.setup_precheck import setup_precheck
 from .setropts.setropts_admin import SetroptsAdmin
 from .user.user_admin import UserAdmin
