+++ tests/common/test_security_request_error.py
@@ -6,7 +6,7 @@ from unittest.mock import Mock, patch
 import __init__
 
 import tests.common.test_common_constants as TestCommonConstants
-from pyracf import SecurityRequestError, UserAdmin
+from pyracf import DownstreamFatalError, UserAdmin
 from pyracf.common.irrsmo00 import IRRSMO00
 
 # Resolves F401
@@ -14,7 +14,7 @@ __init__
 
 
 @patch("pyracf.common.irrsmo00.IRRSMO00.call_racf")
-class TestSecurityRequestError(unittest.TestCase):
+class TestDownstreamFatalError(unittest.TestCase):
     maxDiff = None
     IRRSMO00.__init__ = Mock(return_value=None)
     user_admin = UserAdmin()
@@ -24,7 +24,7 @@ class TestSecurityRequestError(unittest.TestCase):
         call_racf_mock: Mock,
     ):
         call_racf_mock.return_value = [8, 200, 16]
-        with self.assertRaises(SecurityRequestError) as exception:
+        with self.assertRaises(DownstreamFatalError) as exception:
             self.user_admin.set_password("TESTUSER", "Testpass")
         self.assertEqual(
             exception.exception.message,
@@ -37,7 +37,7 @@ class TestSecurityRequestError(unittest.TestCase):
     ):
         call_racf_mock.return_value = [8, 200, 8]
         self.user_admin.set_running_userid("ESWIFT")
-        with self.assertRaises(SecurityRequestError) as exception:
+        with self.assertRaises(DownstreamFatalError) as exception:
             self.user_admin.add("squidwrd")
         self.assertEqual(
             exception.exception.message,
@@ -49,7 +49,7 @@ class TestSecurityRequestError(unittest.TestCase):
         call_racf_mock: Mock,
     ):
         call_racf_mock.return_value = [8, 2000, 20]
-        with self.assertRaises(SecurityRequestError) as exception:
+        with self.assertRaises(DownstreamFatalError) as exception:
             self.user_admin.add("squidwrd")
         self.assertEqual(
             exception.exception.message,
